EquipSet: remove the "ship" attribute
The only way to set it cleanly and transparently was to push
the userdata within the Ship constructor, which caused the promotion
test to fail, leading to death, mayhem, desolation and error screens.

To replace it, the ship is passed as argument wherever it is needed.
I'll wrap those calls in the Ship class, replacing the current API.

A possible beneficial side-effect of this operation might be a slight
performance improvement on all EquipSet actions: table lookups are still
more expensive than direct argument lookup AIUI :-)